open() the output file with proper attributes
Without the mode parameter, open() seems to be creating the output file
with no attributes. This change sets the mode of the output file 0755.

Additionally, at some point during testing it seemed that the output was
not being appended to the file, but instead it was being re-created
every time, keeping the file size at 512 bytes. Because of this, open()
is now being called only once, and the file is closed with close() when
the transfer is done.